More modifications to suit my needs
I found a way to do the verbose logging via i2c_manager instead so I
cleaned up all the hacky code for enabling that. I got super annoyed
with the fact that on macOS VSCode kept giving me red squiggles to
indicate there was an include error (even though there's not really)
just because `__APPLE__` is defined on my system (but not for ESP32) so
I removed that check.

While reading the chip docs I also realized a #define was missing for
`ATCA_CHIP_OPT_AUTO_CLEAR_FAIL` so I added the relevant shift and mask
values for that.